Provide a simple tab separated output formatter
     --output tsv

It will only print the "top level" simple objects of a given result to avoid issues with variable number of columns/row and/or packing complex types into a single column which would require further parsing to make sense of the data.

One can (for now) use --query to project any properties that one wants to expose.