Move command argument modifications out of _autocommand.build_command to application 'command table loaded' event (and exposed 'command table loaded' event from the application). This way, parameter aliasing/providing of metadata works for all commands, not only auto-command commands.
Added validate/morph argument value capabilities to the aliasing subsystem. Previously, one had to use an awkward mix of argparse.Action to populate the namespace attributes and then add an event handler to manipulate the values of the namespace if you wanted to validate combination of arguments or combine argument values from multiple arguments into a single attribute on the namespace (example of the latter is typed $filter parameters)

Currently, network and compute are passing tests. The compute actions (particularly for create scenarios) need to be moved to the new functionality.