keyvault fix for import of pem and pfx (#2754)
Provides better documentation for Key Vault certificate importing showing the path for creating a service principal, importing the cert into key vault, and using it to provision a vm via secrets.

Also fixes an issue where we az keyvault certificate import was not sniffing the type of certificate and properly specifying the content type. Now, we sniff out either PEM or PFX and properly encode the payload for the server.