Randomize server name in test_sql_server_mgmt.
It seems a littly hacky to use SqlServerPreparer just to generate server name, but I couldn't see how to use SingleValueReplacer on its own.

I also ran into issues where the two SqlServerPreparers were generating the same random name - this seemed to be because they were running so quickly (due to not actually creating a resource) that utilities.create_random_name was getting the same timestamp twice in a row. So I added RANDOM_NAME_COUNT into the mix in order to prevent duplicate random names.