Remove marker options from blob, container, and share list commands (#3998)
Issue: https://github.com/Azure/azure-cli/issues/3745

The marker option never works in the CLI context because the commands
never return the marker value when --num-results is specified. A fix is
being worked to improve overall paging mechanism in CLI. In the
meantime, this option is removed to reduce confusing.