Ensure the privates folder's existance in droid (#6333)
The enable build a droid image from privates build, changes are made to
the docker file assuming the existence of the ./privates folder. To
ensure the build success, the folder's existence should be ensured.

This change add a placeholder file to keep this folder exists during the
droid build process.

This is simplier solution, comparing to a if-else structure in the
Docker file.