Cleared up inconsistent language. (#7424)
* Cleared up inconsistent language. 

The short descriptions for ```--service-cidr``` and ```--docker-bridge-address``` are inconsistent. Both require a CIDR notation IP range. However, while the description of ```--service-cidr``` call it a CIDR notation IP range in the ```--docker-bridge-address``` description it is mentioned as an IP address and netmask. I've noticed this creates confusion among less experienced users.

* Updated description to explain the usage better.