Remove our compression of source.
The main reason we needed our own source of compression was to avoid needing to tag a commit in order to get a tarball for upstream_url in our debian and homebrew definitions. Because the GitHub developer API will fetch the appropriate download link for any ref, not just a tag, this isn't actually necessary. This will help us build a candidate homebrew formula without committing to any storage of our own.