Remove Publishing Steps from Build Pipeline (#8178)
* Remove our compression of source.

The main reason we needed our own source of compression was to avoid needing to tag a commit in order to get a tarball for upstream_url in our debian and homebrew definitions. Because the GitHub developer API will fetch the appropriate download link for any ref, not just a tag, this isn't actually necessary. This will help us build a candidate homebrew formula without committing to any storage of our own.

* Removing AzureFileCopy from Build Defintion.

In order to safely allow builds for PRs from forks, we should not commit to any hosting of build products in our public storage account. In that capacity, the AzureFileCopy phases previously defined here should be removed, and the pipeline's permissiosn to do so stripped.

* Explicitly Set Homebrew Version

Before this change, the version was inferred from the url text. When the URL was switched from relying on a tag to an arbirtary ref, we lost the ability to infer the version. This commit looks to re-establish it.