Invoke scripts indirectly using the script task.
This is an alternative to using the Bash@3 task in Azure DevOps. I hypothesize that we are seeing failures because the scripts rely on being executed as a script with their own name from their relative directory. Instead of doing this (as I had expected) Azure DevOps seems to be copying the source of the script into a different file with a UUID name in the root directory and executing. This hampers the scripts ability to use `dirname $0` to figure out relative paths.