Updating ADO branch triggers
It seems like Azure DevOps isn't recognizing that it needs to build the dev and release branches after a PR is merged. Previously, this was moot because there was no difference between the definition of a PR build and a CI build. Now that PR builds do not build the Debian packages, it is necessary to trigger batch builds independently.

This may not be sufficient, and we may need to disable batch builds or continue to trigger release builds manually.