Generate random OAuth state instead of using 'code' as state. (#8512)
* Generate random OAuth state instead of using 'code' as state.

This is a somewhat academic problem, as this is a developer tool running on localhost and not running across a network. However, according to https://tools.ietf.org/html/rfc6749#section-4.1.1 state should be opaque (e.g. hard to guess.) The easiest way for us to accomodate that is to generate a random value for each login request through the CLI and validate that it is what we get back.

* Updating HISTORY.rst

* Fixing PEP8 violation

* Responding to review feedback

* More feedback fixes