Adding existence check to role assignment creation
Fixes #8568

The role assignment creation REST API is not idempotent when the HTTP verb PUT is used. If an existing role is there, instead of returning a 200 it returns a 409 (even though the requests are identical, and definitely not in conflict.) For that reason I've added a list operation first to see if any exist before sending the creation request.

This is a curious thing, as other provisioning tools will have run into the same problem. For instance, how do ARM templates handle this?