Updating Homebrew Formula to use a virtualenv.
This brings us into conformance with the guidelines for how to install Python applications in Homebrew:
https://github.com/Homebrew/brew/blob/5eaf4e5c28e7605cd9d8e1631d376b0e46c98624/docs/Python-for-Formula-Authors.md#installing

However, it is worth noting that after some internal discussions, it seems that there may be more clever ways to achieve the same levels of isolation. This change is not a philosophic disagreement with those ideas, just a pragmatic way forward until a better way is obvious.