Update Homebrew Formula to install in Virtualenv (#8744)
* Updating Homebrew Formula to use a virtualenv.

This brings us into conformance with the guidelines for how to install Python applications in Homebrew:
https://github.com/Homebrew/brew/blob/5eaf4e5c28e7605cd9d8e1631d376b0e46c98624/docs/Python-for-Formula-Authors.md#installing

However, it is worth noting that after some internal discussions, it seems that there may be more clever ways to achieve the same levels of isolation. This change is not a philosophic disagreement with those ideas, just a pragmatic way forward until a better way is obvious.

* Filter out 'jeepney' and 'entrypoints' from homebrew resources

This is a workaround for pypa/pip#6222. It should not be a permanent change.

* Copying wheel-identification regex into our package.

This allows us to remove our dependency on an older version of the 'wheel' package.

Rob Pike has a saying, "A little copy is better than a little dependency". I believe
it's very applicable here, and you can hear him talk about that philosophy here:

https://www.youtube.com/watch?v=PAAkCSZUG1c&t=9m28s

* Moving WHEEL_INFO_RE to __init__

* Fixing PEP8