Migrate Automation Test Runs to ADO (#9359)
* Rough add of pipleline changes

* Fix duplicate Job ID

* Taking advantage of bash to make bearing finding logic more portable.

The previous mechanism for determining the current folder doesn't work in Azure DevOps, only Travis CI. We've learned this the hard way. With this change, it'll work in any bash session, but would not work in old-school shell sessions.

* Upgrade wheel before build

* Fix formatting

* fixing share_folder

* Fix formatting

* Renaming phases to be more appropriate

* Removing Automation Phase from TravisCI

* Remove unnecessary whitespace

A couple of extra newlines were inadvertently added. This commit removes them.