Correct Units in Help and Usage Info
This commit updates the help documentation and usage information to use
the correct IEC unit of measurement for storage (GiB), which should help
reduce any confusion about how these commands work. All storage on Azure
is base 2, and should be using the shorthand for Gibibytes (GiB), not
Gigabytes (GB).

This should bring the documentation for this repository in line with the
storage documentation elsewhere for Azure storage.

rebase