[Compute] Continue parsing image argument if aliases doc unreachable (#10658)
* Continue parsing image argument if aliases doc unreachable

Currently the Azure CLI does not work in environments without public internet access when trying to use the managed disk image in #5 of the _parse_image_argument function. 

This fix ignores the ConnectionError in #4 when trying to lookup the aliases doc from `githubusercontent.com` and proceeds to #5. If #5 does not match then it will return the error message:
```
Failed to connect to remote source of image aliases. Invalid image "{}". Use a valid image URN, customer image name, customer image id, or VHD blob URI.
See vm create -h for more information on specifying an image.
```

* Updated to fix Check CLI Style

* Don't override global ConnectionError