Fix the circular dependency caused by the help find integration
As part of the prototyping for the find integration with the help files
in Cloud Shell, a circular dependency was introduced between the find
modular and the cli-core package. This uses patching to remove that dependency
with the fewest amount of changes possible. it does change the flow of the
project though, so it may need more iterations.