[SQL] BREAKING CHANGE: az sql db create: Remove WideWorldImportersStd and WideWorldImportersFull as allowed values for --sample-name values (#11976)
* Remove WideWorldImporters as allowed values for sql db sample name

The WideWorldImportersStd and WideWorldImportersFull sql db samples were never supported in production and were only documented here by mistake. Removing them from allowed values so that users do not attempt to create them. Note that this is not a breaking change because these options already caused db creation to fail.

* format history

Co-authored-by: Zunli Hu <zuh@microsoft.com>