add required python version in azure-cli setup.py
Currently, pip assumes 2.14.0 is the latest version for all python interpreters and install that version. This causes errors like this on older python versions:

```
Traceback (most recent call last):
  File "<string>", line 1, in <module>
  File "/home/confus/.local/share/virtualenvs/specter-yzp-I12c/lib/python2.7/site-packages/azure/cli/core/commands/__init__.py", line 425
    def cached_put(cmd_obj, operation, parameters, *args, setter_arg_name='parameters', **kwargs):
                                                                        ^
SyntaxError: invalid syntax
```

This probably a systemic problem in more than one azure libraries and because older packages with this error are already on pypi.org, it is hard to solve retrospectively.