Determine plan information when using image alias (#20174)
Some marketplace images require plan information to be provided when creating a
virtual machine. Missing plan information results in the error message:

  "Creating a virtual machine from Marketplace image or a custom image sourced from a Marketplace image requires Plan information in the request."

The "Flatcar" alias is one such case where plan information is needed.  When an
image is provided as a URN, the validator code takes care of determining the
plan information and filling it in, but when using an alias no such action is
taken.

Perform the same lookup when using an alias, so that users can use the alias as
intended and don't have to manually supply plan information.

Signed-off-by: Jeremi Piotrowski <jpiotrowski@microsoft.com>