[Bicep] fix: Use requests environment variables for CA bundle
Azure CLI relies on the requests python package, which
allows users to set custom CA bundle paths through two
environment variables: CURL_CA_BUNDLE and REQUESTS_CA_BUNDLE.

This is useful for cases where the host is behind a MITM proxy
that re-signs all SSL traffic for DPI, or due to a corporate
policy that defines which CA roots are to be trusted.

Enforcing a hard-coded bundle, like the one provided by the certifi
package, only forces users under similar use cases to modify the Python
site packages files, which is meant to be controlled only by the package
manager and thus it is not a good practice to configure the tool.