(AzureCXP) fixes Azure/azure-cli#23097
(AzureCXP) fixes Azure/azure-cli#23097

In this command, there is no documentation that shows where I can get the values for the "name" parameter. So this PR mentions to run the 'az billing period list' command to list the name of billing period' and then use this name in the 'az billing period show' command.