Add test for Backup subtask duration changes.
The feature being tested has been deployed in eastuseuap, and has been
manually tested there. The nature of the CLI side changes are to ensure
that the existing behaviour remains unchanged: hence, this test will
pass now, and should also pass for when the feature is deployed to the
test server regions.