Fix `az network dns zone export` to export all ALIAS records for a particular record set name.
Let's say there are two ALIAS records for the same record set name:

    www 3600 AZURE ALIAS A    /subscriptions/...
    www 3600 AZURE ALIAS AAAA /subscriptions/...

Before this change, export would only emit the last record for any record
set name, in this case the AAAA record, because when it processed the second
record it would incorrectly clear the array containing the first record.

With this change, the array is only initialized when the first record is
processed.