Enforce LF-only line endings in git
Having some files using CRLF makes things extremely messy when
using tarballs on Linux, as the git repository gets automatically
converted, but the tarball stores the original line ending,
so the two do not match anymore and importing new releases fails.

Enforce that LF-only line endings get committed to the repository,
and add instructions for Windows developers on how to configure
their local git client to checkout in CRLF, so that CRLF never
propagate to the repository but remain only in local checkouts.