chore!: do not use default `network_plugin` from the SDK
the CLI should not own defaulting the network
profile properties. Defaulting should be owned
completely by AKS RP, so if a customer does
not set the `network_plugin` field then the CLI
should treat it as `None`.

Today, the default provided from the SDK
is `kubenet` but in the future, the AKS RP
will change this default value when omitted.
This means that other clients (ARM, bicep,
etc.) that do not set `network-plugin` will
get a different behavior than the CLI because
those clients won't set anything but the CLI
will continue to include the default from the
SDK.

Signed-off-by: Tyler Lloyd <tylloyd@microsoft.com>