{Misc} Fix UP022. Use capture_output for subprocess.run.
From the documentation.

If capture_output is true, stdout and stderr will be captured.
When used, the internal Popen object is automatically created
with stdout and stderr both set to PIPE. The stdout and stderr
arguments may not be supplied at the same time as capture_output.
If you wish to capture and combine both streams into one, set stdout to
PIPE and stderr to STDOUT, instead of using capture_output.