{Packaging} Fix crash when another Python is installed on the system
Setting PYTHONPATH is done to avoid conflicts with other installations
of Python on the same machine. When it is set incorrectly Python tends
to crash on startup (typically when loading incompatible ctypes).

There were old scripts in <base>/Scripts, which set PYTHONPATH so that
Python prefers the bundled libs over the system ones. The problem with
these scripts is that they set the wrong path.

But these scripts no longer seem to be used. They were replaced by new
ones in wbin. The problem with those is that PYTHONPATH was not set at
all in them.

Update the scripts to set PYTHONPATH correctly, both for bash and cmd.

Fixes #25067, #28004, #29630, #30517, #30536, #30594, #30700, #30792.