Remove unnecessary network call in `az acr cache` subcommand
`az acr cache create` and `az acr cache update` are used to create and
update cache rules under Azure Container Registry. A _cache rule_ is a
sub resource of a container registry.

Currently, when the above two operations are performed, the code first
makes a GET request to _registry_ resource, only to get the id of the
resource by its name. It needs the registry id in order to create
credential set id which is needed when creating and updating a cache
rule. The credential set id can be determined without making the above
GET request. This GET request is made on the registry resource, which
implies the executing user must have 'Reader' role of the registry. This
could lead to user overprivileging roles, e.g. now they have to give
executing user registry read permissions, in additoinal to cache rule
CRUD permissions.

This commit removed the GET registry request in cache create and cache
update workflow. As a result, a role with only cache rule permissions is
able to execute cache rule create and update. This makes our security
model clearer, and also makes the operation more efficient.