Fixed a cache rule creation bug
With latest az cli 2.70 release we unfortunately introduced a
regression. Creating a cache rule without a credential set would crash.
It is caused by a local variable not declared in the correct scope.

A bug issue has been created to track this issue:
https://github.com/Azure/acr/issues/822

This commit fixed the issue. Now creating cache rules work as expected.