Fix containerapp update using yml file
I think the upgrade to the new API version (https://github.com/Azure/azure-cli/pull/31276) lead to `runningStatus` appearing as a field at the top level when the YAML file passed to `az containerapp update` is run through the deserialiser.

It then caused an error before the command was run:

```
(InvalidRequestContent) The request content was invalid and could not be deserialized: 'Could not find member 'runningStatus' on object of type 'ResourceDefinition'. Path 'runningStatus', line 1, position 224.'.
```

The fix is to add it to the existing list of readonly attributes to remove.