Fix exception in `acr check_health`
In _check_registry_health, get_registry_by_name may return `None` for
the registry. But not all references to the registry check for this
first, resulting in exception when it checks whether ABAC is enabled.
This PR checks whether the registry is non-None before it checks
`registry.role_assignment_mode` for ABAC repository permissions.